(edit) rename getAccessionCodesQueryParam to getParsedAccessionCodes, change the order of if condition for response in getServerSideProps, remove an empty string for searchTerms
