+++ src/pages/search/index.js
@@ -7,7 +7,7 @@ import { TextHighlightContextProvider } from 'contexts/TextHighlightContext'
 import fetchSearch from 'helpers/fetchSearch'
 import formatFacetNames from 'helpers/formatFacetNames'
 import formatFacetQueryParams from 'helpers/formatFacetQueryParams'
-import getAccessionCodesQueryParam from 'helpers/getAccessionCodesQueryParam'
+import getParsedAccessionCodes from 'helpers/getParsedAccessionCodes'
 import getPageNumber from 'helpers/getPageNumber'
 import getSearchQueryForAPI from 'helpers/getSearchQueryForAPI'
 import { Button } from 'components/shared/Button'
@@ -74,9 +74,9 @@ export const Search = ({ query, response }) => {
 
   return (
     <>
-      <PageTitle title={`${search || ''} Results -`} />
+      <PageTitle title={`${search} Results -`} />
       <TextHighlightContextProvider
-        match={[search, ...getAccessionCodesQueryParam(search)]}
+        match={[search, ...getParsedAccessionCodes(search)]}
       >
         <FixedContainer pad={{ horizontal: 'large', bottom: 'large' }}>
           <SearchInfoBanner />
@@ -230,7 +230,7 @@ export const getServerSideProps = async ({ query }) => {
 
   const response = await fetchSearch(queryParams, filterOrders)
 
-  if (response.ok && response) {
+  if (response && response.ok) {
     return {
       props: {
         query: formatFacetQueryParams(
